Refactor ReceiveTimeoutHandler to a separate file.
